users

+++ src/main/java/com/cmlteam/flywise/model/User.java
@@ -1,17 +1,17 @@
 package com.cmlteam.flywise.model;
 
 public class User {
-    private final int id;
+    private final long id;
     private final String firstName;
     private final String lastName;
 
-    public User(int id, String firstName, String lastName) {
+    public User(long id, String firstName, String lastName) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
